+++ tests/test_constraints.py
@@ -7,9 +7,10 @@ import pandas as pd
 from oemof.solph import helpers
 
 from oemof import solph
-from oemof.tabular.constraint_facades import GenericIntegralLimit
+from oemof.tabular.constraint_facades import CONSTRAINT_TYPE_MAP
 from oemof.tabular.facades import (
     BackpressureTurbine,
+    Bev,
     Commodity,
     Conversion,
     Dispatchable,
@@ -100,7 +101,8 @@ class TestConstraints:
 
     def get_om(self):
         return solph.Model(
-            self.energysystem, timeindex=self.energysystem.timeindex
+            self.energysystem,
+            timeindex=self.energysystem.timeindex,
         )
 
     def compare_to_reference_lp(self, ref_filename, my_om=None):
@@ -475,7 +477,8 @@ class TestConstraints:
             output_parameters={"custom_attributes": {"emission_factor": 2.5}},
         )
 
-        emission_constraint = GenericIntegralLimit(
+        emission_constraint = CONSTRAINT_TYPE_MAP["generic_integral_limit"]
+        emission_constraint = emission_constraint(
             name="emission_constraint",
             type="e",
             limit=1000,
@@ -487,4 +490,132 @@ class TestConstraints:
 
         emission_constraint.build_constraint(model)
 
-        self.compare_to_reference_lp("emission_constraint.lp", my_om=model)
+    def test_bev_trio(self):
+        el_bus = solph.Bus("el-bus")
+        el_bus.type = "bus"
+        self.energysystem.add(el_bus)
+
+        indiv_mob = solph.Bus("pkm-bus")
+        indiv_mob.type = "bus"
+        self.energysystem.add(indiv_mob)
+
+        pkm_demand = Load(
+            label="pkm_demand",
+            type="Load",
+            carrier="pkm",
+            bus=indiv_mob,
+            amount=200,  # PKM
+            profile=[0, 1, 0],  # drive consumption
+        )
+
+        self.energysystem.add(pkm_demand)
+
+        bev_v2g = Bev(
+            type="bev",
+            label="BEV-V2G",
+            v2g=True,
+            electricity_bus=el_bus,
+            commodity_bus=indiv_mob,
+            storage_capacity=0,
+            loss_rate=0,  # self discharge of storage
+            charging_power=0,
+            balanced=True,
+            expandable=True,
+            initial_storage_level=0,
+            availability=[1, 1, 1, 1],  # Vehicle availability at charger
+            commodity_conversion_rate=5 / 6,  # Energy to pkm
+            efficiency_mob_electrical=5 / 6,  # Vehicle efficiency per 100km
+            efficiency_mob_v2g=5 / 6,  # V2G charger efficiency
+            efficiency_mob_g2v=5 / 6,  # Charger efficiency
+            efficiency_sto_in=5 / 6,  # Storage charging efficiency
+            efficiency_sto_out=5 / 6,  # Storage discharging efficiency,
+            variable_costs=10,  # Charging costs
+            bev_invest_costs=2,
+            invest_c_rate=60 / 20,  # Capacity/Power
+            fixed_investment_costs=1,
+            lifetime=10,
+        )
+
+        self.energysystem.add(bev_v2g)
+
+        bev_flex = Bev(
+            type="bev",
+            label="BEV-inflex",
+            electricity_bus=el_bus,
+            commodity_bus=indiv_mob,
+            storage_capacity=0,
+            loss_rate=0,  # self discharge of storage
+            charging_power=0,
+            availability=[1, 1, 1, 1],
+            v2g=False,
+            balanced=True,
+            expandable=True,
+            initial_storage_level=0,
+            input_parameters={
+                "fix": [0.89856, 0, 0, 0]
+            },  # fixed relative charging profile
+            output_parameters={
+                "fix": [0.16, 0.08, 0.16, 0.12]
+            },  # fixed relative discharging profile
+            commodity_conversion_rate=5 / 6,  # Energy to pkm
+            efficiency_mob_electrical=5 / 6,  # Vehicle efficiency per 100km
+            efficiency_mob_g2v=5 / 6,  # Charger efficiency
+            efficiency_sto_in=5 / 6,  # Storage charging efficiency
+            efficiency_sto_out=5 / 6,  # Storage discharging efficiency,
+            variable_costs=20,  # Charging costs
+            bev_invest_costs=2,
+            invest_c_rate=60 / 20,  # Capacity/Power
+            fixed_investment_costs=1,
+            lifetime=10,
+        )
+        self.energysystem.add(bev_flex)
+
+        bev_fix = Bev(
+            type="bev",
+            label="BEV-G2V",
+            electricity_bus=el_bus,
+            commodity_bus=indiv_mob,
+            storage_capacity=0,
+            loss_rate=0,  # self discharge of storage
+            charging_power=0,
+            availability=[1, 1, 1, 1],
+            v2g=False,
+            balanced=True,
+            expandable=True,
+            initial_storage_level=0,
+            commodity_conversion_rate=5 / 6,  # Energy to pkm
+            efficiency_mob_electrical=5 / 6,  # Vehicle efficiency per 100km
+            efficiency_mob_g2v=5 / 6,  # Charger efficiency
+            efficiency_sto_in=5 / 6,  # Storage charging efficiency
+            efficiency_sto_out=5 / 6,  # Storage discharging efficiency,
+            variable_costs=4,  # Charging costs
+            bev_invest_costs=2,
+            invest_c_rate=60 / 20,  # Capacity/Power
+            fixed_investment_costs=1,
+            lifetime=10,
+        )
+        self.energysystem.add(bev_fix)
+
+        model = solph.Model(self.energysystem)
+
+        year = self.date_time_index.year.min()
+        mob_share_constraint = CONSTRAINT_TYPE_MAP["bev_share_mob"]
+        mob_share_constraint = mob_share_constraint(
+            name=None,
+            type=None,
+            label="BEV",
+            year=year,
+            share_mob_flex_V2G=0.3,
+            share_mob_inflex=0.2,
+            share_mob_flex_G2V=0.5,
+        )
+        mob_share_constraint.build_constraint(model)
+
+        # these constraints are not mandatory as the energy flow through the
+        # facade is already limited by the in & outflow capactiy of the storage
+        invest_constraint = CONSTRAINT_TYPE_MAP["bev_equal_invest"]
+        invest_constraint = invest_constraint(name=None, type=None, year=year)
+
+        invest_constraint.build_constraint(model)
+
+        self.compare_to_reference_lp("bev_trio_constraint.lp", my_om=model)