+++ src/oemof/tabular/examples/datapackages/dispatch_multi_period_periodic_values/scripts/infer.py
@@ -11,7 +11,7 @@ if "kwargs" not in locals():
     kwargs = {}
 
 building.infer_metadata(
-    package_name="oemof-tabular-dispatch-example",
+    package_name="dispatch_multi_period_periodic_values-example",
     foreign_keys={
         "bus": ["volatile", "dispatchable", "storage", "load"],
         "profile": ["load", "volatile"],