+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 
 Features
 * Improve the function to infer package metadata `#173 <https://github.com/oemof/oemof-tabular/pull/173>`_
+* Add facades `CommodityGHG` and `ConversionGHG` to enable multiple output flows (emissions) `#180 <https://github.com/oemof/oemof-tabular/pull/180>`_
 
 Fixes